Moving p1IterateEcPmSlices files from 1.0.0 one folder up
